% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.sgSEMp1.R
\name{summary.sgSEMp1}
\alias{summary.sgSEMp1}
\title{Summary of Semi-gSEM}
\usage{
\method{summary}{sgSEMp1}(object, ...)
}
\arguments{
\item{object}{An object of class "sgSEMp1", the returned list from sgSEMp1().}

\item{...}{Other arguments. Currently not used.}
}
\value{
NULL. A summary of data and fitting result is printed on screen.
}
\description{
Summarizes the gSEM principle 1 result.
}
\details{
summary.sgSEMp1 gives a summary about the gSEm-Principle 1 analysis.
}
\examples{
data(acrylic)
ans <- sgSEMp1(acrylic)
summary(ans)
}

