% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.sgSEMp2.R
\name{summary.sgSEMp2}
\alias{summary.sgSEMp2}
\title{Summary of Semi-gSEM}
\usage{
\method{summary}{sgSEMp2}(object, ...)
}
\arguments{
\item{object}{An object of class "sgSEMp2", the returned list from sgSEMp2().}

\item{...}{Other arguments. Currently not used.}
}
\value{
NULL. A summary of data and fitting result is printed on screen.
}
\description{
Summarizes the gSEM principle 2 result.
}
\details{
summary.sgSEMp1 gives a brief summary about the gSEM Principle 2 analysis.
}
\examples{
data(acrylic)
ans <- sgSEMp2(acrylic)
summary(ans)
}

